@@ -10,7 +10,7 @@ members = [
 
 [workspace.package]
 edition = "2021"
-rust-version = "1.70"
+rust-version = "1.80"
 license = "MIT OR Apache-2.0"
 homepage = "https://github.com/x-software-com/mxl-crates"
 repository = "https://github.com/x-software-com/mxl-crates"
@@ -25,7 +25,6 @@ tempfile = "3"
 relm4-icons = "0.9"
 # Internationalization:
 i18n-embed-fl = { version = "0.9" }
-once_cell = { version = "1" }
 rust-embed = { version = "8", features = ["debug-embed"] }
 i18n-embed = { version = "0.15", features = [
     "fluent-system",
